pyedgeiotframework/requirements/requirements.txt
sentry-sdk==0.10.2
paho-mqtt>=1.3.1
requests>=2.22.0
python-dateutil>=2.8.0
Pypubsub>=3.3
pyserial>=3.4
python-socketio[client]